requirements.txt
Flask>=2.2
Pillow>=9.0
requests>=2.28
